Basic use of ES6 javascript modules. Needs further splitting and implementation of drawing functions
