+++ targetmine/webapp/src/main/webapp/WEB-INF/web.properties
@@ -301,7 +301,7 @@ head.css.all.targetmineGraphDisplayer = targetmineGraphDisplayer.css
 
 # load custom js for BioActivityGraphDisplayer and GeneExpressionGraphDisplayer
 #  Visualization
-head.js.all.targetmine_d3 = d3-5.9.2-min.js
-head.js.all.targetmine_TargetMineGraph = TargetMineGraph.js
-head.js.all.targetmine_BioActivityGraph = BioActivityGraph.js
+head.js.all.targetmine_A_d3 = d3-5.9.2-min.js
+head.js.all.targetmine_Ba_TargetMineGraph = TargetMineGraph.mjs
+head.js.all.targetmine_Bb_BioActivityGraph = BioActivityGraph.mjs
 # head.js.all.targetmine_GeneExpressionGraph = GeneExpressionGraph.js
